controller: nuke sink implementation
