+++ .golangci.yml
@@ -60,6 +60,10 @@ linters-settings:
         alias: mcsapiv1a1
       - pkg: k8s.io/api/(\w+)/(v[\w\d]+)
         alias: $1$2
+      - pkg: k8s.io/apimachinery/pkg/api/errors
+        alias: apierrors
+      - pkg: github.com/envoyproxy/ai-gateway/internal/testing
+        alias: internaltesting
   gci:
     sections:
       # Captures all standard packages if they do not match another section.
