+++ internal/controller/backend_security_policy.go
@@ -7,13 +7,14 @@ package controller
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"time"
 
 	egv1a1 "github.com/envoyproxy/gateway/api/v1alpha1"
 	"github.com/go-logr/logr"
 	"golang.org/x/oauth2"
-	"k8s.io/apimachinery/pkg/api/errors"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/client-go/kubernetes"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -29,50 +30,49 @@ const preRotationWindow = 5 * time.Minute
 
 // backendSecurityPolicyController implements [reconcile.TypedReconciler] for [aigv1a1.BackendSecurityPolicy].
 //
-// This handles the BackendSecurityPolicy resource and sends it to the config sink so that it can modify configuration.
+// Exported for testing purposes.
 type backendSecurityPolicyController struct {
-	client         client.Client
-	kube           kubernetes.Interface
-	logger         logr.Logger
-	eventChan      chan ConfigSinkEvent
-	oidcTokenCache map[string]*oauth2.Token
+	client               client.Client
+	kube                 kubernetes.Interface
+	logger               logr.Logger
+	oidcTokenCache       map[string]*oauth2.Token
+	syncAIServiceBackend syncAIServiceBackendFn
 }
 
-func newBackendSecurityPolicyController(client client.Client, kube kubernetes.Interface, logger logr.Logger, ch chan ConfigSinkEvent) *backendSecurityPolicyController {
+func newBackendSecurityPolicyController(client client.Client, kube kubernetes.Interface, logger logr.Logger, syncAIServiceBackend syncAIServiceBackendFn) *backendSecurityPolicyController {
 	return &backendSecurityPolicyController{
-		client:         client,
-		kube:           kube,
-		logger:         logger,
-		eventChan:      ch,
-		oidcTokenCache: make(map[string]*oauth2.Token),
+		client:               client,
+		kube:                 kube,
+		logger:               logger,
+		oidcTokenCache:       make(map[string]*oauth2.Token),
+		syncAIServiceBackend: syncAIServiceBackend,
 	}
 }
 
 // Reconcile implements the [reconcile.TypedReconciler] for [aigv1a1.BackendSecurityPolicy].
-func (b *backendSecurityPolicyController) Reconcile(ctx context.Context, req ctrl.Request) (res ctrl.Result, err error) {
+func (c *backendSecurityPolicyController) Reconcile(ctx context.Context, req ctrl.Request) (res ctrl.Result, err error) {
 	var backendSecurityPolicy aigv1a1.BackendSecurityPolicy
-	if err = b.client.Get(ctx, req.NamespacedName, &backendSecurityPolicy); err != nil {
-		if errors.IsNotFound(err) {
-			ctrl.Log.Info("Deleting Backend Security Policy",
+	if err = c.client.Get(ctx, req.NamespacedName, &backendSecurityPolicy); err != nil {
+		if apierrors.IsNotFound(err) {
+			c.logger.Info("Deleting Backend Security Policy",
 				"namespace", req.Namespace, "name", req.Name)
 			return ctrl.Result{}, nil
 		}
 		return ctrl.Result{}, err
 	}
-
 	if oidc := getBackendSecurityPolicyAuthOIDC(backendSecurityPolicy.Spec); oidc != nil {
 		var rotator rotators.Rotator
 		switch backendSecurityPolicy.Spec.Type {
 		case aigv1a1.BackendSecurityPolicyTypeAWSCredentials:
 			region := backendSecurityPolicy.Spec.AWSCredentials.Region
 			roleArn := backendSecurityPolicy.Spec.AWSCredentials.OIDCExchangeToken.AwsRoleArn
-			rotator, err = rotators.NewAWSOIDCRotator(ctx, b.client, nil, b.kube, b.logger, backendSecurityPolicy.Namespace, backendSecurityPolicy.Name, preRotationWindow, roleArn, region)
+			rotator, err = rotators.NewAWSOIDCRotator(ctx, c.client, nil, c.kube, c.logger, backendSecurityPolicy.Namespace, backendSecurityPolicy.Name, preRotationWindow, roleArn, region)
 			if err != nil {
 				return ctrl.Result{}, err
 			}
 		default:
 			err = fmt.Errorf("backend security type %s does not support OIDC token exchange", backendSecurityPolicy.Spec.Type)
-			ctrl.Log.Error(err, "namespace", backendSecurityPolicy.Namespace, "name", backendSecurityPolicy.Name)
+			c.logger.Error(err, "namespace", backendSecurityPolicy.Namespace, "name", backendSecurityPolicy.Name)
 			return ctrl.Result{}, err
 		}
 
@@ -80,12 +80,12 @@ func (b *backendSecurityPolicyController) Reconcile(ctx context.Context, req ctr
 		var rotationTime time.Time
 		rotationTime, err = rotator.GetPreRotationTime(ctx)
 		if err != nil {
-			b.logger.Error(err, "failed to get rotation time, retry in one minute")
+			c.logger.Error(err, "failed to get rotation time, retry in one minute")
 		} else {
 			if rotator.IsExpired(rotationTime) {
-				requeue, err = b.rotateCredential(ctx, &backendSecurityPolicy, *oidc, rotator)
+				requeue, err = c.rotateCredential(ctx, &backendSecurityPolicy, *oidc, rotator)
 				if err != nil {
-					b.logger.Error(err, "failed to rotate OIDC exchange token, retry in one minute")
+					c.logger.Error(err, "failed to rotate OIDC exchange token, retry in one minute")
 				}
 			} else {
 				requeue = time.Until(rotationTime)
@@ -93,24 +93,22 @@ func (b *backendSecurityPolicyController) Reconcile(ctx context.Context, req ctr
 		}
 		res = ctrl.Result{RequeueAfter: requeue}
 	}
-	// Send the backend security policy to the config sink so that it can modify the configuration together with the state of other resources.
-	b.eventChan <- backendSecurityPolicy.DeepCopy()
-	return
+	return res, c.syncBackendSecurityPolicy(ctx, &backendSecurityPolicy)
 }
 
 // rotateCredential rotates the credentials using the access token from OIDC provider and return the requeue time for next rotation.
-func (b *backendSecurityPolicyController) rotateCredential(ctx context.Context, policy *aigv1a1.BackendSecurityPolicy, oidcCreds egv1a1.OIDC, rotator rotators.Rotator) (time.Duration, error) {
+func (c *backendSecurityPolicyController) rotateCredential(ctx context.Context, policy *aigv1a1.BackendSecurityPolicy, oidcCreds egv1a1.OIDC, rotator rotators.Rotator) (time.Duration, error) {
 	bspKey := backendSecurityPolicyKey(policy.Namespace, policy.Name)
 
 	var err error
-	validToken, ok := b.oidcTokenCache[bspKey]
+	validToken, ok := c.oidcTokenCache[bspKey]
 	if !ok || validToken == nil || rotators.IsBufferedTimeExpired(preRotationWindow, validToken.Expiry) {
-		oidcProvider := oauth.NewOIDCProvider(b.client, oidcCreds)
+		oidcProvider := oauth.NewOIDCProvider(c.client, oidcCreds)
 		validToken, err = oidcProvider.FetchToken(ctx)
 		if err != nil {
 			return time.Minute, err
 		}
-		b.oidcTokenCache[bspKey] = validToken
+		c.oidcTokenCache[bspKey] = validToken
 	}
 
 	token := validToken.AccessToken
@@ -142,4 +140,26 @@ func getBackendSecurityPolicyAuthOIDC(spec aigv1a1.BackendSecurityPolicySpec) *e
 // backendSecurityPolicyKey returns the key used for indexing and caching the backendSecurityPolicy.
 func backendSecurityPolicyKey(namespace, name string) string {
 	return fmt.Sprintf("%s.%s", name, namespace)
+}
+
+func (c *backendSecurityPolicyController) syncBackendSecurityPolicy(ctx context.Context, bsp *aigv1a1.BackendSecurityPolicy) error {
+	key := backendSecurityPolicyKey(bsp.Namespace, bsp.Name)
+	var aiServiceBackends aigv1a1.AIServiceBackendList
+	err := c.client.List(ctx, &aiServiceBackends, client.MatchingFields{k8sClientIndexBackendSecurityPolicyToReferencingAIServiceBackend: key})
+	if err != nil {
+		return fmt.Errorf("failed to list AIServiceBackendList: %w", err)
+	}
+
+	var errs []error
+	for i := range aiServiceBackends.Items {
+		aiBackend := &aiServiceBackends.Items[i]
+		c.logger.Info("Syncing AIServiceBackend", "namespace", aiBackend.Namespace, "name", aiBackend.Name)
+		if err = c.syncAIServiceBackend(ctx, aiBackend); err != nil {
+			errs = append(errs, fmt.Errorf("%s/%s: %w", aiBackend.Namespace, aiBackend.Name, err))
+		}
+	}
+	if len(errs) > 0 {
+		return errors.Join(errs...)
+	}
+	return nil
 }
