+++ tests/controller/controller_test.go
@@ -14,6 +14,7 @@ import (
 	"fmt"
 	"log/slog"
 	"os"
+	"sort"
 	"testing"
 	"time"
 
@@ -25,7 +26,6 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/klog/v2"
 	"k8s.io/utils/ptr"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -35,6 +35,7 @@ import (
 
 	aigv1a1 "github.com/envoyproxy/ai-gateway/api/v1alpha1"
 	"github.com/envoyproxy/ai-gateway/internal/controller"
+	internaltesting "github.com/envoyproxy/ai-gateway/internal/testing"
 	testsinternal "github.com/envoyproxy/ai-gateway/tests/internal"
 )
 
@@ -51,16 +52,11 @@ func extProcName(aiGatewayRouteName string) string {
 // TestStartControllers tests the [controller.StartControllers] function.
 func TestStartControllers(t *testing.T) {
 	c, cfg, k := testsinternal.NewEnvTest(t)
-	opts := controller.Options{
-		ExtProcImage:         "envoyproxy/ai-gateway-extproc:foo",
-		EnableLeaderElection: false,
-	}
-	l := logr.FromSlogHandler(slog.NewTextHandler(os.Stderr, &slog.HandlerOptions{}))
-	klog.SetLogger(l)
+	opts := controller.Options{ExtProcImage: "envoyproxy/ai-gateway-extproc:foo", EnableLeaderElection: false}
 
 	ctx := t.Context()
 	go func() {
-		err := controller.StartControllers(ctx, cfg, l, opts)
+		err := controller.StartControllers(ctx, cfg, defaultLogger(), opts)
 		require.NoError(t, err)
 	}()
 
@@ -334,9 +330,8 @@ func TestStartControllers(t *testing.T) {
 
 func TestAIGatewayRouteController(t *testing.T) {
 	c, cfg, k := testsinternal.NewEnvTest(t)
-	ch := make(chan controller.ConfigSinkEvent)
 
-	rc := controller.NewAIGatewayRouteController(c, k, logr.Discard(), ch)
+	rc := controller.NewAIGatewayRouteController(c, k, defaultLogger(), "gcr.io/ai-gateway/extproc:latest", "info")
 
 	opt := ctrl.Options{Scheme: c.Scheme(), LeaderElection: false, Controller: config.Controller{SkipNameValidation: ptr.To(true)}}
 	mgr, err := ctrl.NewManager(cfg, opt)
@@ -345,9 +340,8 @@ func TestAIGatewayRouteController(t *testing.T) {
 	err = ctrl.NewControllerManagedBy(mgr).For(&aigv1a1.AIGatewayRoute{}).Complete(rc)
 	require.NoError(t, err)
 
-	ctx := t.Context()
 	go func() {
-		err := mgr.Start(ctx)
+		err := mgr.Start(t.Context())
 		require.NoError(t, err)
 	}()
 
@@ -389,20 +383,76 @@ func TestAIGatewayRouteController(t *testing.T) {
 			},
 		},
 	}
+
+	for _, b := range []string{"backend1", "backend2"} {
+		err := c.Create(t.Context(), &aigv1a1.AIServiceBackend{
+			ObjectMeta: metav1.ObjectMeta{Name: b, Namespace: "default"},
+			Spec: aigv1a1.AIServiceBackendSpec{
+				APISchema: defaultSchema,
+				BackendRef: gwapiv1.BackendObjectReference{
+					Name: gwapiv1.ObjectName(b),
+					Port: ptr.To[gwapiv1.PortNumber](8080),
+				},
+			},
+		})
+		require.NoError(t, err)
+	}
 	t.Run("create route", func(t *testing.T) {
-		err := c.Create(ctx, origin)
+		err := c.Create(t.Context(), origin)
 		require.NoError(t, err)
 
-		item, ok := <-ch
-		require.True(t, ok)
-		require.IsType(t, &aigv1a1.AIGatewayRoute{}, item)
+		var r aigv1a1.AIGatewayRoute
+		err = c.Get(t.Context(), client.ObjectKey{Name: "myroute", Namespace: "default"}, &r)
+		require.NoError(t, err)
+		require.Equal(t, origin, &r)
 
-		// Verify that they are the same.
-		created := item.(*aigv1a1.AIGatewayRoute)
-		require.Equal(t, "myroute", created.Name)
+		// Verify that the deployment, service, extension policy, and configmap are created.
+		require.Eventually(t, func() bool {
+			deployment, err := k.AppsV1().Deployments("default").Get(t.Context(), extProcName("myroute"), metav1.GetOptions{})
+			if err != nil {
+				t.Logf("failed to get deployment %s: %v", extProcName("myroute"), err)
+				return false
+			}
+			require.Equal(t, "gcr.io/ai-gateway/extproc:latest", deployment.Spec.Template.Spec.Containers[0].Image)
+			require.Len(t, deployment.OwnerReferences, 1)
+			require.Equal(t, origin.Name, deployment.OwnerReferences[0].Name)
+			require.Equal(t, "AIGatewayRoute", deployment.OwnerReferences[0].Kind)
+			require.True(t, *deployment.OwnerReferences[0].Controller)
+			require.Equal(t, int32(5), *deployment.Spec.Replicas)
+			require.Equal(t, resourceReq, &deployment.Spec.Template.Spec.Containers[0].Resources)
 
-		created.TypeMeta = metav1.TypeMeta{} // This will be populated by the controller internally, so we ignore it.
-		require.Equal(t, origin, created)
+			service, err := k.CoreV1().Services("default").Get(t.Context(), extProcName("myroute"), metav1.GetOptions{})
+			if err != nil {
+				t.Logf("failed to get service %s: %v", extProcName("myroute"), err)
+				return false
+			}
+			require.Equal(t, extProcName("myroute"), service.Name)
+			require.Len(t, service.OwnerReferences, 1)
+			require.Equal(t, origin.Name, service.OwnerReferences[0].Name)
+			require.Equal(t, "AIGatewayRoute", service.OwnerReferences[0].Kind)
+			require.True(t, *service.OwnerReferences[0].Controller)
+
+			extPolicy := egv1a1.EnvoyExtensionPolicy{}
+			err = c.Get(t.Context(), client.ObjectKey{Name: extProcName("myroute"), Namespace: "default"}, &extPolicy)
+			if err != nil {
+				t.Logf("failed to get extension policy %s: %v", extProcName("myroute"), err)
+				return false
+			}
+			require.Len(t, extPolicy.OwnerReferences, 1)
+			require.Equal(t, origin.Name, extPolicy.OwnerReferences[0].Name)
+			require.True(t, *extPolicy.OwnerReferences[0].Controller)
+
+			configMap, err := k.CoreV1().ConfigMaps("default").Get(t.Context(), extProcName("myroute"), metav1.GetOptions{})
+			if err != nil {
+				t.Logf("failed to get configmap %s: %v", extProcName("myroute"), err)
+				return false
+			}
+			require.Len(t, configMap.OwnerReferences, 1)
+			require.Equal(t, origin.Name, configMap.OwnerReferences[0].Name)
+			require.True(t, *configMap.OwnerReferences[0].Controller)
+			require.Contains(t, configMap.Data, "extproc-config.yaml")
+			return true
+		}, 30*time.Second, 200*time.Millisecond)
 	})
 
 	t.Run("update", func(t *testing.T) {
@@ -414,42 +464,102 @@ func TestAIGatewayRouteController(t *testing.T) {
 		}
 		origin.Spec.FilterConfig.ExternalProcess.Replicas = ptr.To[int32](3)
 		origin.Spec.FilterConfig.ExternalProcess.Resources = newResource
-		err := c.Update(ctx, origin)
+		err := c.Update(t.Context(), origin)
 		require.NoError(t, err)
 
-		item, ok := <-ch
-		require.True(t, ok)
-		require.IsType(t, &aigv1a1.AIGatewayRoute{}, item)
+		var r aigv1a1.AIGatewayRoute
+		err = c.Get(t.Context(), client.ObjectKey{Name: "myroute", Namespace: "default"}, &r)
+		require.NoError(t, err)
+		require.Equal(t, origin, &r)
 
-		// Verify that they are the same.
-		created := item.(*aigv1a1.AIGatewayRoute)
-		created.TypeMeta = metav1.TypeMeta{} // This will be populated by the controller internally, so we ignore it.
-		require.Equal(t, origin, created)
+		require.Eventually(t, func() bool {
+			deployment, err := k.AppsV1().Deployments("default").Get(t.Context(), extProcName("myroute"), metav1.GetOptions{})
+			if err != nil {
+				t.Logf("failed to get deployment %s: %v", extProcName("myroute"), err)
+				return false
+			}
+			require.Equal(t, "gcr.io/ai-gateway/extproc:latest", deployment.Spec.Template.Spec.Containers[0].Image)
+			require.Len(t, deployment.OwnerReferences, 1)
+			require.Equal(t, origin.Name, deployment.OwnerReferences[0].Name)
+			require.Equal(t, "AIGatewayRoute", deployment.OwnerReferences[0].Kind)
+			require.True(t, *deployment.OwnerReferences[0].Controller)
+			require.Equal(t, int32(3), *deployment.Spec.Replicas)
+			require.Equal(t, newResource, &deployment.Spec.Template.Spec.Containers[0].Resources)
+			return true
+		}, 30*time.Second, 200*time.Millisecond)
 	})
 }
+
+func TestBackendSecurityPolicyController(t *testing.T) { t.Skip("TODO") }
 
 func TestAIServiceBackendController(t *testing.T) {
 	c, cfg, k := testsinternal.NewEnvTest(t)
 
-	ch := make(chan controller.ConfigSinkEvent)
-	bc := controller.NewAIServiceBackendController(c, k, logr.Discard(), ch)
+	syncAIGatewayRoute := internaltesting.NewSyncFnImpl[aigv1a1.AIGatewayRoute]()
 
 	opt := ctrl.Options{Scheme: c.Scheme(), LeaderElection: false, Controller: config.Controller{SkipNameValidation: ptr.To(true)}}
 	mgr, err := ctrl.NewManager(cfg, opt)
 	require.NoError(t, err)
+	require.NoError(t, controller.ApplyIndexing(t.Context(), mgr.GetFieldIndexer().IndexField))
 
+	bc := controller.NewAIServiceBackendController(mgr.GetClient(), k, defaultLogger(), syncAIGatewayRoute.Sync)
 	err = ctrl.NewControllerManagedBy(mgr).For(&aigv1a1.AIServiceBackend{}).Complete(bc)
 	require.NoError(t, err)
 
-	ctx := t.Context()
 	go func() {
-		err := mgr.Start(ctx)
+		err := mgr.Start(t.Context())
 		require.NoError(t, err)
 	}()
 
+	const aiServiceBackendName, aiServiceBackendNamespace = "mybackend", "default"
+	// Create an AIGatewayRoute to be referenced by the AIServiceBackend.
+	originals := []*aigv1a1.AIGatewayRoute{
+		{
+			ObjectMeta: metav1.ObjectMeta{Name: "myroute", Namespace: aiServiceBackendNamespace},
+			Spec: aigv1a1.AIGatewayRouteSpec{
+				APISchema: defaultSchema,
+				TargetRefs: []gwapiv1a2.LocalPolicyTargetReferenceWithSectionName{
+					{
+						LocalPolicyTargetReference: gwapiv1a2.LocalPolicyTargetReference{
+							Name: "gtw", Kind: "Gateway", Group: "gateway.networking.k8s.io",
+						},
+					},
+				},
+				Rules: []aigv1a1.AIGatewayRouteRule{
+					{
+						Matches:     []aigv1a1.AIGatewayRouteRuleMatch{{}},
+						BackendRefs: []aigv1a1.AIGatewayRouteRuleBackendRef{{Name: aiServiceBackendName}},
+					},
+				},
+			},
+		},
+		{
+			ObjectMeta: metav1.ObjectMeta{Name: "myroute2", Namespace: aiServiceBackendNamespace},
+			Spec: aigv1a1.AIGatewayRouteSpec{
+				APISchema: defaultSchema,
+				TargetRefs: []gwapiv1a2.LocalPolicyTargetReferenceWithSectionName{
+					{
+						LocalPolicyTargetReference: gwapiv1a2.LocalPolicyTargetReference{
+							Name: "gtw", Kind: "Gateway", Group: "gateway.networking.k8s.io",
+						},
+					},
+				},
+				Rules: []aigv1a1.AIGatewayRouteRule{
+					{
+						Matches:     []aigv1a1.AIGatewayRouteRuleMatch{{}},
+						BackendRefs: []aigv1a1.AIGatewayRouteRuleBackendRef{{Name: aiServiceBackendName}},
+					},
+				},
+			},
+		},
+	}
+	for _, route := range originals {
+		require.NoError(t, c.Create(t.Context(), route))
+	}
+
 	t.Run("create backend", func(t *testing.T) {
 		origin := &aigv1a1.AIServiceBackend{
-			ObjectMeta: metav1.ObjectMeta{Name: "mybackend", Namespace: "default"},
+			ObjectMeta: metav1.ObjectMeta{Name: aiServiceBackendName, Namespace: aiServiceBackendNamespace},
 			Spec: aigv1a1.AIServiceBackendSpec{
 				APISchema: defaultSchema,
 				BackendRef: gwapiv1.BackendObjectReference{
@@ -458,71 +568,131 @@ func TestAIServiceBackendController(t *testing.T) {
 				},
 			},
 		}
-		err := c.Create(ctx, origin)
+		err := c.Create(t.Context(), origin)
 		require.NoError(t, err)
 
-		item, ok := <-ch
-		require.True(t, ok)
-		require.IsType(t, &aigv1a1.AIServiceBackend{}, item)
+		require.Eventually(t, func() bool {
+			return len(syncAIGatewayRoute.GetItems()) == 2
+		}, 5*time.Second, 200*time.Millisecond)
 
 		// Verify that they are the same.
-		created := item.(*aigv1a1.AIServiceBackend)
-		require.Equal(t, origin, created)
+		routes := syncAIGatewayRoute.GetItems()
+		sort.Slice(routes, func(i, j int) bool {
+			routes[i].TypeMeta = metav1.TypeMeta{}
+			routes[j].TypeMeta = metav1.TypeMeta{}
+			return routes[i].Name < routes[j].Name
+		})
+		require.Equal(t, originals, routes)
+	})
+
+	syncAIGatewayRoute.Reset()
+	t.Run("update backend", func(t *testing.T) {
+		var origin aigv1a1.AIServiceBackend
+		err := c.Get(t.Context(), client.ObjectKey{Name: aiServiceBackendName, Namespace: aiServiceBackendNamespace}, &origin)
+		require.NoError(t, err)
+		origin.Spec.BackendRef.Port = ptr.To[gwapiv1.PortNumber](9090)
+		require.NoError(t, c.Update(t.Context(), &origin))
+
+		require.Eventually(t, func() bool {
+			return len(syncAIGatewayRoute.GetItems()) == 2
+		}, 5*time.Second, 200*time.Millisecond)
+
+		// Verify that they are the same.
+		routes := syncAIGatewayRoute.GetItems()
+		sort.Slice(routes, func(i, j int) bool {
+			routes[i].TypeMeta = metav1.TypeMeta{}
+			routes[j].TypeMeta = metav1.TypeMeta{}
+			return routes[i].Name < routes[j].Name
+		})
+		require.Equal(t, originals, routes)
 	})
 }
 
 func TestSecretController(t *testing.T) {
 	c, cfg, k := testsinternal.NewEnvTest(t)
 
-	ch := make(chan controller.ConfigSinkEvent)
-	sc := controller.NewSecretController(c, k, logr.Discard(), ch)
-
 	opt := ctrl.Options{Scheme: c.Scheme(), LeaderElection: false, Controller: config.Controller{SkipNameValidation: ptr.To(true)}}
 	mgr, err := ctrl.NewManager(cfg, opt)
 	require.NoError(t, err)
 
+	bspSyncFn := internaltesting.NewSyncFnImpl[aigv1a1.BackendSecurityPolicy]()
+	sc := controller.NewSecretController(mgr.GetClient(), k, defaultLogger(), bspSyncFn.Sync)
+	const secretName, secretNamespace = "mysecret", "default"
+
 	err = ctrl.NewControllerManagedBy(mgr).For(&corev1.Secret{}).Complete(sc)
 	require.NoError(t, err)
+	require.NoError(t, controller.ApplyIndexing(t.Context(), mgr.GetFieldIndexer().IndexField))
 
-	ctx := t.Context()
-	go func() {
-		err := mgr.Start(ctx)
-		require.NoError(t, err)
-	}()
+	go func() { require.NoError(t, mgr.Start(t.Context())) }()
+
+	// Create a bsp that references the secret.
+	originals := []*aigv1a1.BackendSecurityPolicy{
+		{
+			ObjectMeta: metav1.ObjectMeta{Name: "mybsp", Namespace: "default"},
+			Spec: aigv1a1.BackendSecurityPolicySpec{
+				Type:   aigv1a1.BackendSecurityPolicyTypeAPIKey,
+				APIKey: &aigv1a1.BackendSecurityPolicyAPIKey{SecretRef: &gwapiv1.SecretObjectReference{Name: secretName}},
+			},
+		},
+		{
+			ObjectMeta: metav1.ObjectMeta{Name: "mybsp2", Namespace: "default"},
+			Spec: aigv1a1.BackendSecurityPolicySpec{
+				Type: aigv1a1.BackendSecurityPolicyTypeAWSCredentials,
+				AWSCredentials: &aigv1a1.BackendSecurityPolicyAWSCredentials{
+					Region:          "us-west-2",
+					CredentialsFile: &aigv1a1.AWSCredentialsFile{SecretRef: &gwapiv1.SecretObjectReference{Name: secretName}},
+				},
+			},
+		},
+	}
+	for _, bsp := range originals {
+		require.NoError(t, c.Create(t.Context(), bsp))
+	}
 
 	t.Run("create secret", func(t *testing.T) {
-		origin := &corev1.Secret{
-			ObjectMeta: metav1.ObjectMeta{Name: "mysecret", Namespace: "default"},
-			StringData: map[string]string{
-				"key": "value",
-			},
-		}
-		err := c.Create(ctx, origin)
+		err := c.Create(t.Context(), &corev1.Secret{
+			ObjectMeta: metav1.ObjectMeta{Name: secretName, Namespace: secretNamespace},
+			StringData: map[string]string{"key": "value"},
+		})
 		require.NoError(t, err)
 
-		item, ok := <-ch
-		require.True(t, ok)
-		event, ok := item.(controller.ConfigSinkEventSecretUpdate)
-		require.True(t, ok)
-		require.Equal(t, "default", event.Namespace)
-		require.Equal(t, "mysecret", event.Name)
+		require.Eventually(t, func() bool {
+			return len(bspSyncFn.GetItems()) == 2
+		}, 5*time.Second, 200*time.Millisecond)
+
+		// Verify that they are the same.
+		bsps := bspSyncFn.GetItems()
+		sort.Slice(bsps, func(i, j int) bool {
+			bsps[i].TypeMeta = metav1.TypeMeta{}
+			bsps[j].TypeMeta = metav1.TypeMeta{}
+			return bsps[i].Name < bsps[j].Name
+		})
+		require.Equal(t, originals, bspSyncFn.GetItems())
 	})
 
+	bspSyncFn.Reset()
 	t.Run("update secret", func(t *testing.T) {
-		origin := &corev1.Secret{
+		err := c.Update(t.Context(), &corev1.Secret{
 			ObjectMeta: metav1.ObjectMeta{Name: "mysecret", Namespace: "default"},
-			StringData: map[string]string{
-				"key": "value2",
-			},
-		}
-		err := c.Update(ctx, origin)
+			StringData: map[string]string{"key": "value2"},
+		})
 		require.NoError(t, err)
 
-		item, ok := <-ch
-		require.True(t, ok)
-		event, ok := item.(controller.ConfigSinkEventSecretUpdate)
-		require.True(t, ok)
-		require.Equal(t, "default", event.Namespace)
-		require.Equal(t, "mysecret", event.Name)
+		require.Eventually(t, func() bool {
+			return len(bspSyncFn.GetItems()) == 2
+		}, 5*time.Second, 200*time.Millisecond)
+
+		bsps := bspSyncFn.GetItems()
+		// Verify that they are the same.
+		sort.Slice(bsps, func(i, j int) bool {
+			bsps[i].TypeMeta = metav1.TypeMeta{}
+			bsps[j].TypeMeta = metav1.TypeMeta{}
+			return bsps[i].Name < bsps[j].Name
+		})
+		require.Equal(t, originals, bsps)
 	})
+}
+
+func defaultLogger() logr.Logger {
+	return logr.FromSlogHandler(slog.NewTextHandler(os.Stderr, &slog.HandlerOptions{}))
 }
